fix(learning): address PR review feedback on candidate observation capture
- `candidate_judgments::upsert_all` counts its own input, so it returns
  `usize` and no caller narrows a database integer with a silent fallback.
- The live-table derivation now requires an identifier's double quotes to
  balance: `CREATE TABLE "foo` yields no table rather than `foo`, so a
  malformed migration fails the real-database comparison instead of hiding
  in it.
- `tests/common/observation_corpus.rs` owns `seeded()` and `QUERY`, which
  both judge suites had defined identically; it includes the git helpers
  itself, so neither suite injects them.
- The config-bound test asserts the field, the env var AND the bound.
- Two schema doc comments no longer overstate what the DDL enforces: the
  `observation_version` column deliberately has no `CHECK` (a downgrade must
  be able to read the row to refuse it by name), and the relevance `CHECK`
  spells its bound literally because SQL cannot read `MAX_RELEVANCE` — now
  tied to the constant by `the_check_bound_is_exactly_max_relevance`.

+++ src/store/candidate_judgments.rs
@@ -33,17 +33,18 @@ pub struct NewJudgment<'a> {
 }
 
 /// Write every verdict in `rows` as one unit, replacing any previous verdict
-/// on the same candidate. Returns how many rows were written.
+/// on the same candidate. Returns how many rows were written — a count of
+/// `rows`, so it is a `usize` and no caller has to narrow a database integer.
 ///
 /// `INSERT OR REPLACE` is correct here and is not the row-preserving-upsert
 /// case the store forbids: the row has no unassigned fields to reset and
 /// nothing references it, so replacing a verdict IS the intended semantics.
-pub fn upsert_all(conn: &Connection, rows: &[NewJudgment<'_>]) -> Result<u64> {
+pub fn upsert_all(conn: &Connection, rows: &[NewJudgment<'_>]) -> Result<usize> {
     if rows.is_empty() {
         return Ok(0);
     }
     let tx = conn.unchecked_transaction()?;
-    let mut written = 0_u64;
+    let mut written = 0_usize;
     for row in rows {
         tx.prepare_cached(
             "INSERT OR REPLACE INTO candidate_judgments\
